refactor(openapi-validator): flatten layout and split configuration
Splits the grab-bag configuration.ts into specifications.ts (schema registry + versions) and errors.ts (error messages), flattens the arbitrary utils/ folder into src/, and renames the misleadingly-named errors.test.ts to error-messages.test.ts.
