+++ packages/openapi-validator/src/index.ts
@@ -1,3 +1,3 @@
-export { OpenApiVersions as supportedVersions } from '@/configuration'
+export { OpenApiVersions as supportedVersions } from '@/specifications'
 export type { ErrorObject, OpenApiVersion, ThrowOnErrorOption, ValidationOutcome } from '@/types'
 export { type ValidateOptions, validate } from '@/validate'
