+++ packages/openapi-validator/src/transform-errors.ts
