+++ packages/openapi-validator/src/validate-path-parameters.test.ts
