+++ packages/openapi-validator/src/validate-path-parameters.ts
